Fix tensor conversions

+++ src/main/scala/io/hydrosphere/sonar/utils/TensorProtoOps.scala
@@ -10,7 +10,7 @@ object TensorProtoOps {
       case DoubleTensor(_, data) => data
       case FloatTensor(_, data) => data.map(_.toDouble)
       case integers @ (_: Uint64Tensor | _: Int64Tensor | _: Int32Tensor | _: Uint32Tensor | _: Uint8Tensor) =>
-        integers.data.asInstanceOf[Seq[Int]].map(_.toDouble)
+        integers.data.asInstanceOf[Seq[Long]].map(_.toDouble)
       case BoolTensor(_, data) => data.map(x => if (x) 1.0 else 0.0)
 
       case unknownTensorType => throw new Exception(s"$unknownTensorType cannot be cast to Seq[Double]")
